Refactored; moved logic inside shadowGetDOMInsidePage to another file
